refactor(maestro): extract task helpers to reduce complexity and duplicates
Extracted findNextEligibleTask, topPendingForManifest, readManifestSafe,
buildNextActionInstruction, and SESSION_MANIFEST_DIR into new next-task.js
module. Factored reconcileTask() and aliveTicketSet() out of syncFromTmux
in manifest.js to lower cyclomatic complexity below 10 and eliminate
jscpd duplicate-blocks. Actions.js now stays under 400-line limit.
Fixes Quality gate CI failures from PR #533.
